Add importer for Passman (#13)
* Add importer for Passman

The importer reads JSON files generated by the Nextcloud app Passman.

The first tag is used as the folder name. This works well if passwords
have no or only one tag. If no username is set then the email address is
used as the username.

Files are not being imported.

* Fix indentation

* Remove unsettable revision date

* Fix tslint findings

* Add email to notes when there is also a username

If a username is set, that will become the username. Otherwise the email
will become the username.

If a username and an email is set the email will be added to the notes.